backend/pyproject.toml
[tool.poetry]
name = "backend"
package-mode = false
version = "0.1.0"
description = "FastAPI Platform API"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.128.7"
uvicorn = "^0.40.0"
python-dotenv = "^1.0.1"
livekit-api = "^1.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.46"}
asyncpg = "^0.31.0"
alembic = "^1.18.4"
redis = {extras = ["hiredis"], version = "^7.1.1"}
bcrypt = "4.0.1"
pydantic-settings = "^2.12.0"
python-json-logger = "^4.0.0"
pyjwt = "^2.11.0"
email-validator = "^2.3.0"
passlib = "^1.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)",
    "ruff (>=0.15.0,<0.16.0)",
    "pre-commit (>=4.5.1,<5.0.0)"
]
[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["backend"]
